+++ tests/misc/tests/misc.ts
@@ -1,18 +1,20 @@
-const anchor = require("@project-serum/anchor");
-const assert = require("assert");
-const {
-  ASSOCIATED_TOKEN_PROGRAM_ID,
+import * as anchor from "@project-serum/anchor";
+import { Program, BN, IdlAccounts, AnchorError } from "@project-serum/anchor";
+import {
+  PublicKey,
+  Keypair,
+  SystemProgram,
+  SYSVAR_RENT_PUBKEY,
+} from "@solana/web3.js";
+import {
   TOKEN_PROGRAM_ID,
   Token,
-} = require("@solana/spl-token");
-const miscIdl = require("../target/idl/misc.json");
-const {
-  SystemProgram,
-  Keypair,
-  PublicKey,
-  SYSVAR_RENT_PUBKEY,
-} = require("@solana/web3.js");
+  ASSOCIATED_TOKEN_PROGRAM_ID,
+} from "@solana/spl-token";
+import { Misc } from "../target/types/misc";
 const utf8 = anchor.utils.bytes.utf8;
+const assert = require("assert");
+const miscIdl = require("../target/idl/misc.json");
 
 describe("misc", () => {
   // Configure the client to use the local cluster.
@@ -231,9 +233,8 @@ describe("misc", () => {
       assert.ok(false);
     } catch (err) {
       const errMsg = "A close constraint was violated";
-      assert.equal(err.toString(), errMsg);
-      assert.equal(err.msg, errMsg);
-      assert.equal(err.code, 2011);
+      assert.equal(err.error.errorMessage, errMsg);
+      assert.equal(err.error.errorCode.number, 2011);
     }
   });
 
@@ -702,7 +703,7 @@ describe("misc", () => {
           });
         },
         (err) => {
-          assert.equal(err.code, 2009);
+          assert.equal(err.error.errorCode.number, 2009);
           return true;
         }
       );
@@ -737,7 +738,7 @@ describe("misc", () => {
           });
         },
         (err) => {
-          assert.equal(err.code, 2015);
+          assert.equal(err.error.errorCode.number, 2015);
           return true;
         }
       );
@@ -872,7 +873,7 @@ describe("misc", () => {
         },
       }),
       (err) => {
-        assert.equal(err.code, 2006);
+        assert.equal(err.error.errorCode.number, 2006);
         return true;
       }
     );
@@ -990,8 +991,10 @@ describe("misc", () => {
         },
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 2004);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2004);
     }
   });
 
@@ -1027,8 +1030,10 @@ describe("misc", () => {
         },
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 2006);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2006);
     }
   });
 
@@ -1054,8 +1059,10 @@ describe("misc", () => {
         signers: [newAcc],
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 2019);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2019);
     }
   });
 
@@ -1086,8 +1093,10 @@ describe("misc", () => {
         signers: [mint],
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 2016);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2016);
     }
   });
 
@@ -1118,8 +1127,10 @@ describe("misc", () => {
         signers: [mint],
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 2017);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2017);
     }
   });
 
@@ -1150,8 +1161,10 @@ describe("misc", () => {
         signers: [mint],
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 2018);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2018);
     }
   });
 
@@ -1195,8 +1208,10 @@ describe("misc", () => {
         signers: [token],
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 2015);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2015);
     }
   });
 
@@ -1252,8 +1267,10 @@ describe("misc", () => {
         signers: [token],
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 2014);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2014);
     }
   });
 
@@ -1303,8 +1320,10 @@ describe("misc", () => {
         },
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 2015);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2015);
     }
   });
 
@@ -1366,8 +1385,10 @@ describe("misc", () => {
         },
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 2014);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2014);
     }
   });
 
@@ -1430,8 +1451,10 @@ describe("misc", () => {
         },
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 3014);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 3014);
     }
   });
 
@@ -1456,8 +1479,10 @@ describe("misc", () => {
         signers: [data],
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(err.code, 2005);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2005);
     }
   });
 
@@ -1582,9 +1607,14 @@ describe("misc", () => {
         },
       });
       assert.ok(false);
-    } catch (err) {
-      assert.equal(2005, err.code);
-      assert.equal("A rent exempt constraint was violated", err.msg);
+    } catch (_err) {
+      assert.ok(_err instanceof AnchorError);
+      const err: AnchorError = _err;
+      assert.equal(err.error.errorCode.number, 2005);
+      assert.equal(
+        "A rent exemption constraint was violated",
+        err.error.errorMessage
+      );
     }
   });
 
@@ -1611,8 +1641,10 @@ describe("misc", () => {
           },
         });
         assert.ok(false);
-      } catch (err) {
-        assert.equal(err.code, 2006);
+      } catch (_err) {
+        assert.ok(_err instanceof AnchorError);
+        const err: AnchorError = _err;
+        assert.equal(err.error.errorCode.number, 2006);
       }
 
       // matching bump seed for wrong address but derived from wrong program
@@ -1624,8 +1656,10 @@ describe("misc", () => {
           },
         });
         assert.ok(false);
-      } catch (err) {
-        assert.equal(err.code, 2006);
+      } catch (_err) {
+        assert.ok(_err instanceof AnchorError);
+        const err: AnchorError = _err;
+        assert.equal(err.error.errorCode.number, 2006);
       }
 
       // correct inputs should lead to successful tx
@@ -1661,8 +1695,10 @@ describe("misc", () => {
           },
         });
         assert.ok(false);
-      } catch (err) {
-        assert.equal(err.code, 2006);
+      } catch (_err) {
+        assert.ok(_err instanceof AnchorError);
+        const err: AnchorError = _err;
+        assert.equal(err.error.errorCode.number, 2006);
       }
 
       // same seeds but derived from wrong program
@@ -1674,8 +1710,10 @@ describe("misc", () => {
           },
         });
         assert.ok(false);
-      } catch (err) {
-        assert.equal(err.code, 2006);
+      } catch (_err) {
+        assert.ok(_err instanceof AnchorError);
+        const err: AnchorError = _err;
+        assert.equal(err.error.errorCode.number, 2006);
       }
 
       // correct inputs should lead to successful tx